requirements.txt
autopep8==1.7.0
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-QRcode==3.1.0
Flask-WTF==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
json2html==1.3.0
MarkupSafe==2.1.1
Pillow==9.2.0
pycodestyle==2.9.1
pylightning==0.0.7.3
python-dotenv==0.21.0
qrcode==7.3.1
toml==0.10.2
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
